gvc gvprintpointflist: take number of points as a size_t
This is a more appropriate type for the length of an array.

+++ lib/gvc/gvdevice.c
@@ -595,10 +595,9 @@ void gvprintpointf(GVJ_t * job, pointf p)
     gvwrite(job, buf, len);
 } 
 
-void gvprintpointflist(GVJ_t * job, pointf *p, int n)
-{
+void gvprintpointflist(GVJ_t *job, pointf *p, size_t n) {
   const char *separator = "";
-  for (int i = 0; i < n; ++i) {
+  for (size_t i = 0; i < n; ++i) {
     gvputs(job, separator);
     gvprintpointf(job, p[i]);
     separator = " ";
